@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
         <LangVersion>8</LangVersion>
     </PropertyGroup>
 
@@ -11,9 +11,9 @@
 
     <ItemGroup>
       <PackageReference Include="Microsoft.AspNetCore.Authentication" Version="2.2.0" />
-      <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="5.5.0" />
-      <PackageReference Include="Serilog" Version="2.8.0" />
-      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.5.0" />
+      <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="5.6.0" />
+      <PackageReference Include="Serilog" Version="2.9.0" />
+      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="5.6.0" />
     </ItemGroup>
 
 </Project>
