@@ -3,28 +3,33 @@ using System.Linq;
 
 namespace OpenDND.Services.Core
 {
+    public enum MaximumDieValues
+    {
+        D2 = 2,
+        D4 = 4,
+        D6 = 6,
+        D8 = 8,
+        D10 = 10,
+        D12 = 12,
+        D20 = 20,
+        D100 = 100
+    }
+    
     public class RollDiceService
     {
         private static Random rand;
-        private static object randLock;
-        private readonly int[] validNumbers = { 4, 6, 8, 10, 12, 20, 100 };
 
         public RollDiceService()
         {
             rand = new Random();
-            randLock = new object();
         }
-
-        public int Roll(int maxValue)
+        
+        // Todo: Create methods for rolling single die rolls and rolling multiple dies at the same time
+        public int Roll(MaximumDieValues die)
         {
-            var check = validNumbers.Any(x => x == maxValue);
-
-            if (!check)
-            {
-                throw new ArgumentException($"{maxValue} is not valid");
-            }
-
-            return rand.Next(1, maxValue + 1);
+            var max = (int) die;
+            var roll = rand.Next(1, max + 1);
+            return roll;
         }
     }
 }
