@@ -1,7 +1,9 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using OpenDND.Data.Models.Core;
+using Microsoft.IdentityModel.Tokens;
+using OpenDND.Data.Models.App;
 
 namespace OpenDND.Services.Core
 {
@@ -14,9 +16,9 @@ namespace OpenDND.Services.Core
     public class UserService : IUserService
     {
         // users hardcoded for simplicity, store in a db with hashed passwords in production applications
-        private List<User> _users = new List<User>
+        private readonly List<User> _users = new List<User>
         { 
-            new User { Id = 1, FirstName = "Test", LastName = "User", Username = "test", Password = "test" } 
+            new User(1, "Test", "Test", "Admin", "Pass", Role.Admin, String.Empty)
         };
 
         public async Task<User> Authenticate(string username, string password)
