@@ -1,4 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
+using OpenDND.DiscordBot;
 
 namespace OpenDND.Extensions
 {
@@ -11,7 +12,8 @@ namespace OpenDND.Extensions
         /// <returns></returns>
         public static IServiceCollection AddOpenDNDServices(this IServiceCollection services)
         {
-            
+            services.AddSingleton<DiscordSerilogAdapter>();
+            services.AddHostedService<OpenDndBot>();
             
             return services;
         }
